test get_proxy is integration test.
